Make sure that PR partitions don't exceed integer range for number of nodes (#370)
* partiton pagerank so that every partition holds nodes in the integer range
* use int[] instead of IntArray since partitions are now guaranteed to be < max_int

* use double[] instead of DoubleArray since partitions are now guaranteed to be < max_int

+++ core/src/main/java/org/neo4j/graphalgo/core/utils/HugeParallelExporter.java
@@ -108,7 +108,7 @@ public abstract class HugeParallelExporter<T> extends AbstractExporter<T> {
 
     private void writeParallel(T data) {
         verify();
-        final int batchSize = Math.min(
+        final int batchSize = (int) Math.min(
                 MAX_BATCH_SIZE,
                 ParallelUtil.adjustBatchSize(
                         nodeCount,
